==> 0a_Downsample_All.ipynb <==
#import standard libs
import numpy as np
import pandas as pd
from tqdm import tqdm
import importlib

#utils
import mt_utils as ut

importlib.reload(ut)

# claspy dataloader libs
from claspy.data_loader import load_tssb_dataset
from claspy.data_loader import load_has_dataset

#load benchmarks
tssb = load_tssb_dataset()
hasc = load_has_dataset()

#result definitions
columns = ['Orig_TS_ID', 'DS_Algo', 'CF', 'DS_TS_index']

n_tssb = len(tssb)
n_hasc = len(hasc)
algos = ["MinMax","MinMaxLTTB","M4","LTTB", "EveryNth","LTD"]

#compression ratios
crs = np.round(np.arange(0.1,1,0.1),1)

# TSSB

results_tssb = []
for index, row in tqdm(tssb.iterrows(), total=len(tssb), desc="Processing rows"):
    ts = row.time_series
    # downsample ts using all algos and all crs
    for algo in algos:
        for cr in crs:
            n_out = ut.adjust_nout(algo, int(cr*len(ts)))
            ds_ts_i = ut.downsample(algo, ts, n_out)
            
            results_tssb.append((index, algo, cr, ds_ts_i))

results_df_TSSB = pd.DataFrame(results_tssb, columns=columns)
results_df_TSSB.to_pickle("results/all_downsampled_TSSB.pkl")
results_df_TSSB

# HASC

results_hasc = []
for index, row in tqdm(hasc.iterrows(), total=len(hasc), desc="Processing rows"):
    ts_arr = np.transpose(row.time_series)
    for algo in algos:
        for cr in crs:
            ds_ts_list = []
            for ts in ts_arr:
                n_out = ut.adjust_nout(algo, int(cr*len(ts)))
                ds_ts_i = ut.downsample(algo, ts, n_out)
                ds_ts_list.append((ds_ts_i))
            ds_ts_i = np.transpose(np.array(ds_ts_list))
            
            results_hasc.append((index, algo, cr, ds_ts_i))

results_df_HASC = pd.DataFrame(results_hasc, columns=columns)
results_df_HASC.to_pickle("results/all_downsampled_HASC.pkl")
results_df_HASC
