==> 3a_Window_Size.ipynb <==
#import standard libs
import numpy as np
import pandas as pd

# load previous datasets
tssb_res_dsW = pd.read_pickle("results/zwischenergebnisse/clasp_on_downsampled_TS_TSSB_dsW_ucps.pkl")
tssb_res_origW = pd.read_pickle("results/zwischenergebnisse/clasp_on_downsampled_TS_TSSB_origW_ucps.pkl")

hasc_res_dsW = pd.read_pickle("results/zwischenergebnisse/clasp_on_downsampled_TS_HASC_dsW_ucps.pkl")
hasc_res_origW = pd.read_pickle("results/zwischenergebnisse/clasp_on_downsampled_TS_HASC_origW_ucps.pkl")

def score_comparison_matrix_w(df):
    mean = df.groupby(by=['method']).mean().rename(columns={'score': 'mean'})
    median = df.groupby(by=['method']).median().rename(columns={'score': 'median'})
    std = df.groupby(by=['method']).std().rename(columns={'score': 'std'})
    
    out = mean.merge(median, left_on='method', right_on='method')
    out = out.merge(std, left_on='method', right_on='method')
    out = out.transpose()
    return out

# use default cr 0.5
cr = 0.5

# TSSB

tssb_res_dsW_filtered = tssb_res_dsW[tssb_res_dsW['compression']==cr].reset_index(drop=True)
tssb_res_origW_filtered = tssb_res_origW[tssb_res_origW['compression']==cr].reset_index(drop=True)

# filter samples
tssb_res_dsW_filtered['method']='Downscaled SuSS'
tssb_res_origW_filtered['method']='SuSS on downsampled TS'

tssb_res_all = pd.concat([tssb_res_dsW_filtered,tssb_res_origW_filtered],axis=0).reset_index(drop=True)
# compare mean scores and check which one is better
window_sizes_SumMat_TSSB = score_comparison_matrix_w(tssb_res_all[['method','score']])
window_sizes_SumMat_TSSB.to_excel("results/tables/summary_statistics_window_size_approaches_TSSB.xlsx")
window_sizes_SumMat_TSSB

# HASC

hasc_res_dsW_filtered =hasc_res_dsW[hasc_res_dsW['compression']==cr].reset_index(drop=True)
hasc_res_origW_filtered = hasc_res_origW[hasc_res_origW['compression']==cr].reset_index(drop=True)

hasc_res_dsW_filtered['method']='Downscaled SuSS'
hasc_res_origW_filtered['method']='SuSS on downsampled TS'

hasc_res_all = pd.concat([hasc_res_dsW_filtered,hasc_res_origW_filtered],axis=0).reset_index(drop=True)
# compare mean scores and check which one is better
window_sizes_SumMat_HASC = score_comparison_matrix_w(hasc_res_all[['method','score']])
window_sizes_SumMat_HASC.to_excel("results/tables/summary_statistics_window_size_approaches_HASC.xlsx")
window_sizes_SumMat_HASC
